breaking change: JSON objects are now alists and JSON arrays vectors
After discussions in the guile-user mailing list, using types with reader
syntax seemed the best way to represent JSON types for most people.

Most of the time people build JSON objects so it seems logic to map them to
the most convenient way to represent them in Scheme which would be alists. For
JSON arrays we now use vectors which also have reader syntax #() and therefore
are easy to represent.